// src/types/Psychologist.ts
export interface Psychologist {
  id: string;
  email: string;
  nome: string;
  crp: string;
  regiao: string;
  instagram: string;
  numeroWhatsapp: string;
  dataNascimento: string;
  cpf: string;
  formacoes: string[];
  experiencias: string[];
  anoGraduacao: number;
  abordagem: string;
  valorConsulta: number;
  imageUrl: string;
}